Add `enabled_languages` option (#3)
* Add `enabled_languages` option

Defaults to all languages being opted in to being checked, but with this
option can specify specific languages to opt into.

Also, expands the tests, bumps the Go version, and does a bit of
cleanup.

Resolves #1.

* Fix option conversion

We want options to be supplied as lowercase, but the languages
themselves can stay regular cased based on their proper names for errors
and such. (I guess eventually we'll need a mapping, but lowercasing is
simple enough for now.)